RemoteFragment.java: all standard buttons now use ontouch + keyup/dn
instead of onclick + keypress. This allows for a smoother control for
situations where holding a button down is handled (as a test case,
check the Retaliate Roku Game, or scrolling through videos on YouTube
and moving through a video replay).

Even the main Roku menu uses rew/ff buttons as PgUp/PgDn alternatives
to skip options by screenfulls, and changing the behaviour of the
buttons here has an impact too. Same case for back in standard
navigation.

So, this is the only behaviour for buttons that emmit key events now.