VibratingImageButton: added touch listener support
to allow for haptic feedback just as the button is touched instead of
at the end on the click process, and to allow for the click process to
be prevented so that finer keyup and keydown commands can be sent to
the Roku while preventing a duplicitous keypress to be sent at the end
of the interaction. A click/keypress can still be used in case the
button is "clicked" by different means than touch input.

VIBRATE_DURATION_MS reduced to 25ms, and provideHapticFeedback calls
were replaced with provideHapticEffect with EFFECT_TICK for a more
subtle vibration pattern.